+++ projects/fabric/src/main/kotlin/site/siredvin/peripheralium/FabricPeripheralium.kt
@@ -5,6 +5,7 @@ import net.fabricmc.fabric.api.event.lifecycle.v1.ServerWorldEvents
 import site.siredvin.peripheralium.fabric.FabricIngredients
 import site.siredvin.peripheralium.fabric.FabricLibPlatform
 import site.siredvin.peripheralium.fabric.FabricPeripheraliumPlatform
+import site.siredvin.peripheralium.fabric.FabricXplatTags
 import site.siredvin.peripheralium.storages.FabricStorageUtils
 import site.siredvin.peripheralium.storages.fluid.FluidStorageExtractor
 import site.siredvin.peripheralium.storages.item.ItemStorageExtractor
@@ -13,7 +14,7 @@ import site.siredvin.peripheralium.xplat.LibCommonHooks
 object FabricPeripheralium : ModInitializer {
 
     init {
-        PeripheraliumCore.configure(FabricLibPlatform, FabricPeripheraliumPlatform, FabricIngredients)
+        PeripheraliumCore.configure(FabricLibPlatform, FabricPeripheraliumPlatform, FabricIngredients, FabricXplatTags)
         // Register extract storages
         ItemStorageExtractor.addStorageExtractor(FabricStorageUtils::extractStorage)
         FluidStorageExtractor.addFluidStorageExtractor(FabricStorageUtils::extractFluidStorage)