+++ projects/fabric/src/test/java/site/siredvin/peripheralium/tests/PeripheraliumInitialization.java
@@ -6,11 +6,12 @@ import site.siredvin.peripheralium.PeripheraliumCore;
 import site.siredvin.peripheralium.fabric.FabricIngredients;
 import site.siredvin.peripheralium.fabric.FabricLibPlatform;
 import site.siredvin.peripheralium.fabric.FabricPeripheraliumPlatform;
+import site.siredvin.peripheralium.fabric.FabricXplatTags;
 
 @AutoService(Extension.class)
 public class PeripheraliumInitialization implements Extension{
 
     public PeripheraliumInitialization() {
-        PeripheraliumCore.INSTANCE.configure(FabricLibPlatform.INSTANCE, FabricPeripheraliumPlatform.INSTANCE, FabricIngredients.INSTANCE);
+        PeripheraliumCore.INSTANCE.configure(FabricLibPlatform.INSTANCE, FabricPeripheraliumPlatform.INSTANCE, FabricIngredients.INSTANCE, FabricXplatTags.INSTANCE);
     }
 }